refactor: extract dashboard assets
